requirements.txt
fastapi
uvicorn
qdrant-client
pypdf
python-dotenv
google-generativeai
sentence-transformers
python-multipart